Fix Windows 32 again

+++ test/test-mc-reader.c
@@ -165,7 +165,7 @@ _test_mc_reader_bytes (_mongocrypt_tester_t *tester)
    ASSERT_OK_STATUS (mc_reader_read_buffer_to_end (&reader, &value_buf, status),
                      status);
    ASSERT_CMPBYTES (
-      expected_bytes + 4u, expected_len - 4u, value_buf.data, value_buf.len);
+      expected_bytes + 4u, (size_t)expected_len - 4u, value_buf.data, value_buf.len);
 
    _mongocrypt_buffer_cleanup (&input_buf);
    _mongocrypt_buffer_cleanup (&value_buf);
